pyproject.toml
[tool.poetry]
name = "keplermatik_test"
version = "0.1.0"
description = ""
authors = ["mysteriousham <mysteriousham73@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rel = "^0.4.7"
numpy = "^1.23.5"
pillow = "^9.3.0"
textual = "^0.7.0"
websocket-client = "^1.4.2"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
